Omaha #7888 Thrift deserialization - fix integer truncation check
Allow deserialized integer values to be assigned to a field whose type is a
superclass of that integer type (namely Number or Object) without checking
for truncation.

Fixes spurious "java.lang.Long cannot be cast to java.lang.Number" error that
would occur when deserializing a Long value into a Number field.

Change-Id: I4dcab9c7b87b78dfa467182e886c868a232ce9ac