Omaha #7724 Qpid API cleanup
New API for getting Qpid connections is to instantiate a
QpidUFConnectionFactory, which is configured by passing in a JMSConnectionInfo
object. Connection strings, SSL configuration, and broker REST API calls are
all encapsulated in the connection factory implementation.
QpidUFConnectionFactory instances should typically be handled by client code as
a ConnectionFactory, i.e. not dependent on the specific JMS implementation.

Change-Id: I1816984e42d41171f6a8be549f35bbad623556d8