Omaha #8326 Rewrite ClusteredQuartzComponent for Camel 3 / Quartz 2
APIs have changed dramatically with the Quartz 2 and Camel 3 upgrades,
necessitating a rewrite of the ClusteredQuartz component.

Specifically, the tactic of creating a modified QuartzEndpoint that overrides
onJobExecute to acquire the cluster lock doesn't work anymore, because
QuartzEndpoint works completely differently now and onJobExecute is gone.

Quartz provides trigger listener functionality that allows one to listen for
the firing of a trigger and "veto" the execution of the job if certain
conditions are met. We can use this functionality to acquire the cluster lock
and if that fails, veto the job.

Change-Id: If76867d30f776d4cc13a297d09161033327eff97