Omaha #8717 Split EDEX FOSS into individual packages
One feature per FOSS project comes out as one RPM package per feature.
Updated the com.raytheon.uf.edex.foss feature to depend on all of these
other features.

All of the new features include the version number of the FOSS package
so that eventually this version number can be included in the RPM
version number. (Extra changes to the build are needed to make that
happen and this commit does not include those changes.)

This commit addresses only those FOSS that were being installed as part
of various awips2-edex-* RPMs. The rest of the EDEX FOSS is in
awips2-common-base, which is a special case that cannot be split up
at the feature level without large changes to the build. I split
awips2-common-base at the RPM level in a separate commit.

Change-Id: I2b4a59e0a768e412cd3e0956db79f7f7507e207c