Omaha #8854 D2D menu performance improvements
Updating menu item text sometimes causes bad performance. Mainly because
whenever new data comes in for a bundle, the menu text for that bundle gets
updated even if the displayed data time did not change. Also whenever the menu
text for a bundle gets updated it triggers an update of all other items in the
menu which massively multiplies the number of UI jobs that get put on queue.

Through testing I found that it's not really necessary for an update of one
menu item to trigger all others getting updated. The data times in the menu are
appearing correctly aligned without this code so I removed it.

I also added code to check if the displayed data time would actually change for
a given bundle before updating the menu text for that bundle. Menu text will
still get refreshed whenever the menu is opened, but not, for example when CAVE
gets 500 data notifications for surface obs that all have the same data time.

Overall this cuts down the menu text updates by a huge amount especially at
times when there is a lot of data coming in.

Change-Id: I46b1915980925fd5daab79016fd55c9496b69cf5