Issue #2037227 New EDEX API for Camel routes
New Java-based API to allow plugins to add Camel routes to EDEX. This
replaces the current paradigm of defining CamelContexts in XML files,
which is no longer possible as of Camel 4.

This commit just includes the new API and the bare minimum necessary
changes to core EDEX code to be able to start EDEX with no plugins.
Routes will start and stop in an unknown order, cluster-locking is
disabled, and context state processors will not run. More work is
needed to address all of those issues.

Change-Id: I26740a43853ccacc0632f3fe05270f75199a90db