Issue #2037700 Camel 4 - Remove now-useless code in ContextData
With the move to one single CamelContext, the ContextData class has been
made nearly obsolete. There is no need to track many CamelContexts and
figure out which is the correct one to send a particular message in,
so most code from that class can now be removed.

However, we still need to keep the routeID-to-endpointURI map for two
reasons:

1. Guarantee uniqueness of route IDs. If you add a route to the Camel
   Context more than once with the same route ID, the newer route
   definition will replace the old. We don't want that. Previously we
   didn't need to worry about it so much because there were many
   CamelContexts and each had its own namespace for route IDs.

2. Throughput. We want to send messages as fast as possible. Even
   though there are not many routes (hundreds), it may still be too
   slow to iterate over all routes whenever sending a message via
   MessageProducer, so we keep the old behavior which is to use a map.

I moved that map to ContextManager where it makes more sense.

MessageProducer can also be cleaned up further (there is still some
code that is too complex because it assumes multiple CamelContexts) but
that can be done later.

Two tangential changes are also included:

1. I made a change to the URI parsing pattern in ContextData because
   it could not handle valid URIs without "//"
   (example: "direct:watchWarnDirect")

2. I temporarily set number of data storage audit queues to zero
   because multiple CamelContexts were still being created for those.
   Once the DataStorageAuditContextsBuilder has been fixed to create
   EDEXRouteContexts instead of CamelContexts then I will undo this.

Change-Id: I17707879d2f4ff8d306c073eebdeca6d656f4ff0