Issue #2037700 Startup/shutdown for EDEXRouteContexts
Replace startup/shutdown of the CamelContext with operating on
EDEXRouteContexts.

These are the minimal changes required to adapt the existing code to
work with EDEXRouteContexts.

There are still problems, especially at shutdown. The order is not
correct. EDEX takes too long to shut down and gets killed by SystemD.
Also, clustered contexts are still not handled.

More work to follow.

Change-Id: I65b4f0d46d53c6a5f631d27140151bbdf19119bf