Issue #2037700 Faster EDEX shutdown (Camel 4)
Because Camel uses internally a single-thread executor to run each
shutdown task, the shutdown of the routes is not parallelized. So it's
still slow. But I was able to make it a bit faster by stopping all timer
endpoints before shutdown and by submitting a shutdown task for each
whole route context rather than each individual route.

Also:
- Deleted the dummy mode since we don't need it now. Request mode works
  well enough for testing purposes.
- Gave names to some threads that should have names for debugging
  purposes.
- Fixed some old typos.

Change-Id: I7525f1665d5fc46afe17ad80099f984829166d4d