Issue #2037843 use PythonEval so shared modules initialize properly
With Python 3.11.9, when opening the GFE perspective in CAVE, CAVE crashes.
This was not seen with Python 3.11.7.

This was eventually traced to the fact that the threading module is
automatically imported in the Python interpreter in 3.11.9 and before
it was not. When GFE starts it creates a SubInterpreter in PyCacheUtil
that does not have any shared modules declared. This is normally ok
because that interpreter doesn't import anything other than sys. However,
when a Jep SubInterpreter is closed, it checks if the threading module was
imported or not and if so then checks an active thread count. This check
for an active thread count causes Jep to go farther down its list of type
checks and then initialize numpy so it can check for numpy arrays. Numpy is
initialized and then shortly thereafter disposed as part of the
SubInterpreter close, and the dispose of numpy causes numpy's state to get
messed up so the next time numpy is used it causes a SIGSEGV and crash.

Change-Id: I3717edd785db206147c861a7851500261838b54c