Issue #2037700 Camel 4 fix internal route shutdown order
Sorting routes by only if they are internal or not is not a total
ordering, so using a reversing comparator does not work as expected
for reversing the order of the list. Instead do the simpler thing
which is to sort the list of routes and then reverse it, this
guarantees that shutdown order will always be exact opposite of
startup order.

Change-Id: I4b4214b24643da4c808a39161aa9021d4da747bd