Issue #2037700 Fix audit queue instantiation for Camel 4
The Camel Context has to be injected into the ContextManager via Spring
and ContextManager in turn has to be Spring-injected into the
DataStorageAuditContextsBuilder, otherwise Spring is not aware of the
dependency relationships and this can result in trying to create
audit queue routes before the CamelContext exists, which is impossible
and fails EDEX startup.

This is potentially a problem in any case where an EDEXRouteBuilder is
created dynamically in Java code rather than in Spring XML. So, don't
do that. Fortunately we only have the one class
(DataStorageAuditContextBuilder) that does that.

Change-Id: I4abb4ae8a96cb664126a390e8921f10003522a66