Issue #2037700 Allow multiple JMS consumer threads for same topic
Background: The DedicatedThreadJmsComponent class creates a separate
thread pool for each Camel JMS endpoint and limits the number of threads
to the number of concurrentConsumers specified in the endpoint URI
(default = 1). Camel uses one shared endpoint for all routes within a
CamelContext that are subscribed to a JMS topic, so that means one
thread per topic per CamelContext.

Now: Since everything is in only one CamelContext, we have to remove the
limit on the number of threads when the endpoint is a topic, to allow an
arbitrary number of subscribers to a particular topic within the same
JVM.

This fixes the problem seen after Camel 4 upgrade where EDEX in
ingestGrib or bmh mode will frequently spew warnings about "paused
tasks" associated with particular JMS endpoints; it turns out that these
two EDEX modes have multiple routes that listen to the same topic.

Change-Id: Ida443058ece538cc051ebf60515165273a1bfc4d