Issue #2037700 Lower Camel JMS receive timeout to 50ms at EDEX shutdown
Speeds up EDEX shutdown by reducing the amount of time spent waiting for
graceful shutdown of each JMS consumer route.

The reason this is necessary is that Camel shuts down one route at a
time within a given CamelContext. It was fine before Camel 4 because the
many CamelContexts themselves could be shut down in parallel. Now with
all the routes in one CamelContext, serial shutdown is way too slow when
Camel has to wait average of 2.5 seconds (half of the JMS receive
timeout during normal EDEX operation) for each JMS consumer route to
shut down.

Maybe it's possible to parallelize shutdown at the level of individual
routes by implementing a custom Camel ShutdownStrategy, but that is
about 100x more complicated than this solution.

Change-Id: I4aa82d185266de890891f988ae9ca3d75148b6cc