Issue #2039687 fix ConcurrentModificationExceptions that happen in RequestLogger
Some IRequestHandlers that RequestLogger is configured to write to json make modifications to their RequestConstraint maps during processing.  This in turn can cause RequestLogger to get a ConcurrentModificationException as it is looping over the map in Jackson to transform the request to json.  Fixed that by making copies of the RequestConstraint maps in the handlers so only the copy map is modified, not the original.

Change-Id: I4645c1e6b4962febf844d7a5b53fc4dc9ab84265